Experiment: JS values as primitive forth values
This is just a snapshot for now, since it barely gets to the "ok"
prompt. The idea is that all JavaScript values become valid primitive
values for Forth (or a Forth-like language). The lack of integers just
makes things more interesting (I've also decided to forego byte
addressing, and ideally make c@ and @ the same thing).

We can't even print numbers right now, though.